requirements.txt
Django>=5.0
djangorestframework
psycopg2-binary
django-import-export
gunicorn
python-dotenv

